pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ovp"
version = "0.1.0"
description = "Open-Vocabulary Perception: GroundingDINO + SAM 2 + ByteTrack pipeline for text-prompted detection, segmentation, and tracking."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Emin Çağan Apaydın", email = "your.email@example.com" },
]
keywords = [
    "computer-vision",
    "object-detection",
    "segmentation",
    "tracking",
    "open-vocabulary",
    "grounding-dino",
    "sam2",
    "bytetrack",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

# NOTE: PyTorch is intentionally NOT listed here.
# Install it manually first, matching your CUDA version:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# Then run: pip install -e .
dependencies = [
    "transformers>=4.45.0,<5.0.0",
    "accelerate>=0.34.0",
    "supervision>=0.25.0",
    "opencv-python>=4.10.0",
    "pillow>=10.0.0",
    "torchvision",
    "numpy>=1.26.0,<2.0.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.8.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "huggingface-hub>=0.25.0",
    "einops>=0.8.0",
    "scipy>=1.13.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "pycocotools>=2.0.7",     # COCO loading
    "torchmetrics>=1.0.0",    # mAP / mIoU
]
notebook = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.1.0",
    "matplotlib>=3.9.0",
]
all = [
    "ovp[dev,notebook]",
]

[project.urls]
Repository = "https://github.com/e-cagan/open-vocab-perception"
Issues = "https://github.com/e-cagan/open-vocab-perception/issues"

[project.scripts]
ovp-image = "ovp.scripts.run_image:app"
ovp-video = "ovp.scripts.run_video:app"
ovp-benchmark = "ovp.scripts.benchmark:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ovp = ["py.typed"]

# ----------------------------- Ruff -----------------------------
[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",     # line too long (handled by formatter)
    "B008",     # function calls in argument defaults (common in DI)
    "RUF003",   # Turkish characters in comments are intentional
    "RUF022",   # __all__ is grouped semantically, not alphabetically
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # allow re-exports
"notebooks/*" = ["E402", "F401"]
"tests/*" = ["B011"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ----------------------------- Mypy -----------------------------
[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
files = ["src"]

# ----------------------------- Pytest -----------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU (deselect with '-m \"not gpu\"')",
]